refactor: sd as 18 decimals (#312)
* refactor: sd as 18 decimals

test: update tests accordingly

* feedback on helpers

* refactor: add "scaled" in ongoing and snapshot debt

test: update tests accordingly

* docs: improve natspec

test: use `ONE_MONTH_DEBT_18D` constant


* last fixes


---------

+++ tests/integration/fuzz/Fuzz.t.sol
@@ -55,7 +55,7 @@ abstract contract Shared_Integration_Fuzz_Test is Integration_Test {
             uint128 amountSeed = uint128(uint256(keccak256(abi.encodePacked(flow.nextStreamId(), decimals))));
             // Bound the amount between a realistic range.
             uint128 amount = boundUint128(amountSeed, 1e18, 200_000e18);
-            uint128 depositAmount = getDescaledAmount(amount, decimals);
+            uint128 depositAmount = uint128(getDescaledAmount(amount, decimals));
 
             // Deposit into the stream.
             deposit(streamId, depositAmount);
